requirements.txt
# Web Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1

# LINE Bot SDK
line-bot-sdk==3.19.0

# Database
aiosqlite==0.21.0
sqlalchemy==2.0.43
greenlet==3.2.4

# LLM Client (OpenAI-compatible API for vLLM)
openai==2.3.0

# Google Maps
googlemaps==4.10.0

# Utilities
requests==2.32.5
pyyaml==6.0.3

# AI (optional - for direct Google Gemini integration)
google-genai==1.36.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
